reuse DataNodeActions component
